fix: 이니셜 메소드 수정
